% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alkohol}
\alias{alkohol}
\title{Alcohol consumption by country}
\format{
A data frame with 12 rows and 3 variables:
\describe{
  \item{Land}{country}
  \item{Typ}{type of alcohol}
  \item{Portionen}{quantity of portions}
}
}
\usage{
alkohol
}
\description{
The dataset was created from data available at \url{https://www.prescient.pro/}.
}
\keyword{datasets}
