% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{autos}
\alias{autos}
\title{Used cars in the USA crawled 2016}
\format{
A data frame with 47972 rows and 23 variables:
\describe{
  \item{DatumCrawled}{date of crawling}
  \item{Name}{name}
  \item{Verkäufer}{seller}
  \item{AngebotsTyp}{type of offer}
  \item{PreisChar}{price as character}
  \item{Test}{test}
  \item{Typ}{type of car}
  \item{Jahr}{year of production}
  \item{Getriebe}{transmission}
  \item{PS}{horsepower}
  \item{Modell}{model}
  \item{Kilometer}{odometer in km}
  \item{Monat}{month of production}
  \item{Motor}{motor}
  \item{Marke}{brand}
  \item{Unfall}{collision}
  \item{Datum}{date}
  \item{AnzahlBilder}{quantity of pictures}
  \item{ZipCode}{postal code}
  \item{DatumSichtung}{date}
  \item{Alter}{age}
  \item{Preis}{price}
  \item{Kategorie}{category}
}
}
\usage{
autos
}
\description{
The dataset was created from data available at \url{https://www.prescient.pro/}.
}
\keyword{datasets}
