% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bundesliga}
\alias{bundesliga}
\title{German soccer league standing 2021}
\format{
A data frame with 18 rows and 12 variables:
\describe{
  \item{R.}{rank}
  \item{V.}{rank previous week}
  \item{Var.3}{not used}
  \item{Verein}{name}
  \item{Sp.}{number of games played}
  \item{S.}{games win}
  \item{U.}{games neiter lost nor won}
  \item{N.}{games lost}
  \item{Tore.}{goals}
  \item{TD.}{difference of goals made versus received}
  \item{P..}{points}
  \item{Quali.}{qualification for champions league}
}
}
\usage{
bundesliga
}
\description{
The dataset was created from data available at \url{https://www.prescient.pro/}.
}
\keyword{datasets}
