% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/classfuntion.R
\name{classfunction}
\alias{classfunction}
\title{Creating a  bivariate normal distribution}
\usage{
classfunction(nx1, nx2, nx3, mux1, mux2, mux3, sdx1, sdx2, sdx3, rhox1, rhox2,
  rhox3, name)
}
\arguments{
\item{nx1}{Number of observations from the class/group 1.}

\item{nx2}{Number of observations from the class/group 2.}

\item{nx3}{Number of observations from the class/group 3.}

\item{mux1}{Vector of two means (c(mean1, mean2)) from each feature from the class/group 1.}

\item{mux2}{Vector of two means (c(mean1, mean2)) from each feature from the class/group 2.}

\item{mux3}{Vector of two means (c(mean1, mean2)) from each feature from the class/group 3.}

\item{sdx1}{Vector of two standard deviations (c(sdx1, sdx2)) from each feature from the class/group 1.}

\item{sdx2}{Vector of two standard deviations (c(sdx1, sdx2)) from each feature from the class/group 2.}

\item{sdx3}{Vector of two standard deviations (c(sdx1, sdx2)) from each feature from the class/group 3.}

\item{rhox1}{Correlation coefficient of the two features in class 1.}

\item{rhox2}{Correlation coefficient of the two features in class 2.}

\item{rhox3}{Correlation coefficient of the two features in class 3.}

\item{name:}{By default, since we are dealing with a three class function, it assigns Dog, Cat and Rabbits name. We can specify it.}
}
\value{
It returns a data frame with the first two columns being the observations from each feature normally distributed, the third column states the group those observations belong to and the last column assigns 1, 2 and 3 to each group.
}
\description{
Creating a  bivariate normal distribution
}
\examples{
# we can create a matrix of bivariate normal distributions with three different classes
class<-classfunction(nx1=200, nx2=200, nx3=200, mux1=c(5,6), mux2=c(1,6), mux3=c(3,4), sdx1=c(1.3,1.3), sdx2=c(0.9,1), sdx3=c(1.2,1),rhox1=-0.72, rhox2=0.3, rhox3=-0.2, name=c("Cocodrile", "Rabbit", "Fish") )
plot(class$Var1, class$Var2, col=class$Label, xlab="Feature 1", ylab="Feature 2", main="Features by groups distribution")
}

