% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colour.own.ampel}
\alias{colour.own.ampel}
\title{Colours for traffic lights}
\format{
Vector with 3 elements with the colour code
\describe{
}
}
\usage{
colour.own.ampel
}
\description{
The theme defines the colours green, yellow and red.
}
\keyword{datasets}
