% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colour.own.nomin.1}
\alias{colour.own.nomin.1}
\title{Colours for nominal variables - Option 1}
\format{
Vector with 11 elements with the colour code
\describe{
}
}
\usage{
colour.own.nomin.1
}
\description{
The theme defines colours for nominal variables.
}
\keyword{datasets}
