% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colour.own.nomin.2}
\alias{colour.own.nomin.2}
\title{Colours for nominal variables - Option 2}
\format{
Vector with 8 elements with the colour code
\describe{
}
}
\usage{
colour.own.nomin.2
}
\description{
The theme defines colours for nominal variables.
}
\keyword{datasets}
