% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colour.own.seque.1}
\alias{colour.own.seque.1}
\title{Colours for sequential variables - Option 1}
\format{
Vector with 9 elements with the colour code
\describe{
}
}
\usage{
colour.own.seque.1
}
\description{
The theme defines colours for sequential variables.
}
\keyword{datasets}
