% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colour.own.seque.2}
\alias{colour.own.seque.2}
\title{Colours for sequential variables - Option 2}
\format{
Vector with 11 elements with the colour code
\describe{
}
}
\usage{
colour.own.seque.2
}
\description{
The theme defines colours for sequential variables.
}
\keyword{datasets}
