% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colour.own.spect.2}
\alias{colour.own.spect.2}
\title{Colours for spectral variables - Option 2}
\format{
Vector with 11 elements with the colour code
\describe{
}
}
\usage{
colour.own.spect.2
}
\description{
The theme defines colours for spectral variables.
}
\keyword{datasets}
