% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df}
\alias{df}
\title{Group of friends with attributes}
\format{
A data frame with 5 rows and 5 variables:
\describe{
  \item{freund}{name of friend}
  \item{fb}{facebook friends}
  \item{alter}{age}
  \item{farbe}{colour favourite}
  \item{altersgruppe}{age group}
}
}
\usage{
df
}
\description{
The dataset was created fictitiously.
}
\keyword{datasets}
