% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/discriminant.R
\name{discriminant}
\alias{discriminant}
\title{Creating a linear discriminant function with 2 features and 3 classes}
\usage{
discriminant(class)
}
\arguments{
\item{class}{It is the matrix we generate. It must contain 2 features and 3 classes}
}
\value{
It returns the slopes and the intercept of the linear discriminant with 2 features and 3 classes.
}
\description{
Creating a linear discriminant function with 2 features and 3 classes
}
\examples{
# we can create a matrix of bivariate normal distributions with three different classes
class1<-classfunction(nx1=600, nx2=600, nx3=600, mux1=c(14,10), mux2=c(4,4), mux3=c(12,0.5), sdx1=c(1.3,1.3), sdx2=c(0.9,1), sdx3=c(1.2,1), rhox1=-0.72, rhox2=0.3, rhox3=-0.2, name=c("Rabbit", "Cocodrile", "Fish"))
discriminant(class1)$intercept
discriminant(class1)$slope
}

