% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{durchschnittsnoten}
\alias{durchschnittsnoten}
\title{Average student grades by gender and age}
\format{
A data frame with 49999 rows and 5 variables
\describe{
  \item{Note}{grade}
  \item{Geschlecht}{gender}
  \item{Alter}{age}
}
}
\usage{
durchschnittsnoten
}
\description{
The dataset was created was created fictitiously.
}
\keyword{datasets}
