\name{EGATUR}
\alias{EGATUR}
\docType{data}
\title{
egatur dataset
}
\description{egatur}
\usage{data("EGATUR")}
\format{
  A data frame with 30541 observations on the following 13 variables.
  \describe{
    \item{\code{mm_aaaa}}{a numeric vector}
    \item{\code{pais}}{a factor with country names }
    \item{\code{A13}}{a numeric vector}
    \item{\code{aloja}}{a factor with levels \code{Hotels} \code{Rest of market} \code{Over-The-Counter Accommodation}}
    \item{\code{motivo}}{a factor with levels \code{Leisure} \code{Business} \code{Others}}
    \item{\code{gastototal}}{a numeric vector}
    \item{\code{factoregatur}}{a numeric vector}
    \item{\code{GastoTotalD}}{a factor with levels \code{[17,1.67e+03)} \code{[1.67e+03,5.51e+03)} \code{[5.51e+03,1.99e+04]}}
    \item{\code{A13_D}}{a factor with levels \code{[1,3)} \code{[3,4)} \code{[4,6)} \code{[6,7)} \code{[7,14)} \code{[14,180]}}
  }
}
\details{
 If necessary, more details than the __description__ above ~~
}
\source{
reference to a publication or URL from which the data were obtained ~~
}
\references{
Test
}
\examples{
data(EGATUR)
## maybe str(EGATUR) ; plot(EGATUR) ...
}
\keyword{datasets}
