% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flug}
\alias{flug}
\title{Flights from New York airports in 2013}
\format{
A data frame with 328521 rows and 28 variables:
\describe{
  \item{dest}{destination}
  \item{year}{year}
  \item{arr_month}{arrival month}
  \item{arr_day}{arrival day}
  \item{arr_hour}{arrival hour}
  \item{origin}{origin}
  \item{month}{origin month}
  \item{day}{origin day}
  \item{hour}{origin hour}
  \item{tailnum}{plane tailnumber ID}
  \item{carrier}{carrier}
  \item{flight}{flight number}
  \item{air_time}{air_time}
  \item{dep_delay}{departure delay}
  \item{arr_delay}{arrival delay}
  \item{origin_alt}{origin altitude}
  \item{origin_tzone}{origin time zone}
  \item{dest_alt}{destination altitude}
  \item{dest_tzone}{destination time zone}
  \item{name}{carrier name}
  \item{age}{age of plane}
  \item{seats}{seats on plane}
  \item{origin_precip}{origin precipitation}
  \item{origin_visib}{origin visibility}
  \item{dest_precip}{destination precipitation}
  \item{dest_visib}{destination visibility}
  \item{quarter}{quarter of year}
  \item{agegroup}{age of plane}
}
}
\usage{
flug
}
\description{
The dataset was created from data available at \url{https://www.prescient.pro/}.
}
\keyword{datasets}
