% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{formel1}
\alias{formel1}
\title{Racing results from formula1 in 2019}
\format{
A data frame with 420 rows and 9 variables:
\describe{
  \item{Rennen}{race number}
  \item{Ort}{place}
  \item{Startposition}{starting order}
  \item{Rennwagen}{race car number}
  \item{Fahrer}{driver}
  \item{Team}{team}
  \item{Platzierung}{result}
  \item{Punkte}{points}
  \item{WRPunkte}{points towards world championship}
}
}
\usage{
formel1
}
\description{
The dataset was created from data available at \url{https://www.prescient.pro/}.
}
\keyword{datasets}
