% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kfz}
\alias{kfz}
\title{Used cars in the USA crawled 2016}
\format{
A data frame with 49999 rows and 5 variables
\describe{
  \item{Marke}{brand}
  \item{Zulassung}{year of first use}
  \item{PS}{horsepower}
  \item{Kilometerstand}{odometer}
  \item{Preis}{price}
}
}
\usage{
kfz
}
\description{
The dataset was created from data available at \url{https://www.prescient.pro/}.
}
\keyword{datasets}
