% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leipzig}
\alias{leipzig}
\title{List of words with sentiment value}
\format{
A data frame with 36 rows and 4 variables:
\describe{
  \item{token}{word}
  \item{anzahl}{count of word in text}
  \item{wert}{sentiment value of word}
  \item{kumwert}{cumulated sentiment value of words}
}
}
\usage{
leipzig
}
\description{
A list of words in a text with their sentiment value based on the Leipzig dictionary \url{https://www.prescient.pro/}.
}
\keyword{datasets}
