% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.eval.kl.r
\name{ml.eval.kl}
\alias{ml.eval.kl}
\title{Function to evaluate the quality of a prediction for categorical variables (classification)}
\usage{
ml.eval.kl(df, titel = "", xvar = "", yvar = "", scatter = F)
}
\arguments{
\item{df}{is a dataframe with two variables: real values, estimated values}

\item{titel}{is the name of the variable to be used as subtitle on diagrams}

\item{xvar}{name of the first variable}

\item{yvar}{name of the second variable}

\item{scatter}{if TRUE a scatter-chart is created}
}
\value{
auc
}
\description{
Function to evaluate the quality of a prediction for categorical variables (classification)
}
\keyword{classification}
\keyword{evaluation,}
\keyword{learning,}
\keyword{machine}
\keyword{modeling,}
