% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.eval.lr.r
\name{ml.eval.lr}
\alias{ml.eval.lr}
\title{Function to evaluate the quality of a prediction for numeric variables (regression)}
\usage{
ml.eval.lr(df, titel = "", xvar = "", yvar = "", scatter = F)
}
\arguments{
\item{df}{is a dataframe with two variables: reak values, estimated values}

\item{titel}{is the name of the variable to be used as subtitle on diagrams}

\item{xvar}{name of the first variable}

\item{yvar}{name of the second variable}

\item{scatter}{if TRUE a scatter-chart is created}
}
\value{
rmse & rsq
}
\description{
Function to evaluate the quality of a prediction for numeric variables (regression)
}
\keyword{evaluation,}
\keyword{learning,}
\keyword{machine}
\keyword{modeling,}
\keyword{regression}
