% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.eval.r
\name{ml.eval}
\alias{ml.eval}
\title{Function to evaluate a machine learning model the correlation for numeric variables}
\usage{
ml.eval(modell, daten, ziel.spalte)
}
\arguments{
\item{modell}{is the name of a machine learning model}

\item{daten}{is a dataframe used for prediction}

\item{ziel.spalte}{is the column number of the variable to be predicted}
}
\value{
model name, dataframe name, rmse & rsq
}
\description{
Function to evaluate a machine learning model the correlation for numeric variables
}
\keyword{evaluation}
\keyword{learning,}
\keyword{machine}
\keyword{modeling,}
