% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.summary.2.r
\name{ml.summary.2}
\alias{ml.summary.2}
\title{Function to calculate descriptive statistics for two variables}
\usage{
ml.summary.2(
  df,
  titel,
  xvar,
  yvar,
  npos = 0,
  dec = 0,
  bw = F,
  hist = F,
  bar = F,
  line = F,
  box = F,
  scatter = F,
  kor = F
)
}
\arguments{
\item{df}{is a dataframe with two variables}

\item{titel}{is the name of the variable to be used as subtitle on diagrams}

\item{xvar}{name of the first variable, a numeric variable}

\item{yvar}{name of the second variable}

\item{npos}{defines the x-value at which the number of items per category is displayed in a box-plot}

\item{dec}{defines the number of decimals behind the period}

\item{bw}{uses only black and white in a box-plot}

\item{hist}{if TRUE a histogram is created}

\item{bar}{if TRUE a bar-chart is created}

\item{line}{if TRUE a line-chart is created}

\item{box}{if TRUE a box-plot is created}

\item{scatter}{if TRUE a scatter-chart is created}

\item{kor}{if TRUE a correlation matrix is created}
}
\value{
text with feedback to the function
}
\description{
Function to calculate descriptive statistics for two variables
}
\keyword{correlation,}
\keyword{descriptive,}
\keyword{plot}
\keyword{statistics,}
