% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.summary.r
\name{ml.summary}
\alias{ml.summary}
\title{Function to calculate descriptive statistics for a variable}
\usage{
ml.summary(variable, titel, achse, bar = F, barminmax = 0, box = F, rel = F)
}
\arguments{
\item{variable}{is a vector of the variable}

\item{titel}{is the name of the variable to be used as title on diagrams}

\item{achse}{is the description of the axis on diagrams}

\item{bar}{if TRUE a bar-chart is created}

\item{barminmax}{defines how many bars shall be displayed}

\item{box}{if TRUE a box-plot is created}

\item{rel}{if TRUE a histogram with the relative values in & is created}
}
\value{
a vector with a list of descriptive statistics
}
\description{
Function to calculate descriptive statistics for a variable
}
\keyword{descriptive,}
\keyword{statistics,}
\keyword{summary}
