% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.tunepar.r
\name{ml.tunepar}
\alias{ml.tunepar}
\title{Function to show the tuning parameters for a workflow or workflow set}
\usage{
ml.tunepar(wfl, wfl.name = "Workflowname")
}
\arguments{
\item{wfl}{is the name of a workflow or workflow set}

\item{wfl.name}{is the name of the workflow or workflow set as title}
}
\value{
a text with a message
}
\description{
Function to show the tuning parameters for a workflow or workflow set
}
\keyword{learning}
\keyword{machine}
\keyword{parameter,}
\keyword{tuning,}
