% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLA.class-functions.R
\name{plot.MLA}
\alias{plot.MLA}
\title{Plot MLA object}
\usage{
\method{plot}{MLA}(x, simply = FALSE, ...)
}
\arguments{
\item{x}{object of class "MLA"}

\item{simply}{Allow to simplify the view of nodes, in case of MLA object is a CART. Default value is FALSE.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function plot an MLA object. It is a method for the generic function plot.
}
