% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLA.class-functions.R
\name{plotCART}
\alias{plotCART}
\title{Plot rpart or MLA object}
\usage{
plotCART(x, ownlabs = TRUE)
}
\arguments{
\item{x}{object of class "MLA" or "rpart".}

\item{ownlabs}{Allow to simplify the view of nodes. Default value is TRUE.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function plot an MLA object or a rpart.
}
