% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLA.class-functions.R
\name{print.MLA}
\alias{print.MLA}
\title{Print an MLA Object}
\usage{
\method{print}{MLA}(x, first = 100, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{object of class "MLA"}

\item{first}{When the print command shows a set of rules limits the number of rules.}

\item{digits}{minimal number of significant digits.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function prints an MLA object. It is a method for the generic function print.
}
