% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler}
\alias{sampler}
\title{Splitting your dataset in training and testing}
\usage{
sampler(data, p, seed = NULL)
}
\arguments{
\item{data}{the data frame that contains the variables to be separeted.}

\item{p}{the percentage of the training dataset to be obtained randomly. It can be expressed in either decimal fraction (such as 0.7) or percent (such as 72.12).}

\item{seed}{a single value, interpreted as an integer, or \code{NULL}. The default value is \code{NULL}, but for future checks of the model or models generated it is advisable to set a random seed to be able to reproduce it.}
}
\description{
A training/test partition are created by \code{sampler} function.
}
\examples{
# The best way to demostrate the functionality is test the function
Sampling <- sampler(EGATUR,p=0.7)



}
