% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{studierende}
\alias{studierende}
\title{Students data from 2020}
\format{
A data frame with 357 rows and 22 variables
\describe{
  \item{Gruppe}{group}
  \item{Geschlecht}{gender}
  \item{Geburtsjahr}{year of birth}
  \item{Geburtsmonat}{month of birth}
  \item{Größe}{height}
  \item{Einwohner}{citicens of city of origin}
  \item{Distanz}{distance of residence to university}
  \item{DistanzEltern}{distance of residence of parents to university}
  \item{Buchstaben}{length of first name}
  \item{Fußball}{like or dislike of soccer}
  \item{Schuhe}{pairs of shoes in posession}
  \item{Intelligenz}{intelligence}
  \item{Attraktivität}{attractiveness}
  \item{Lieblingsfarbe}{preferred colour}
  \item{Hobby}{hobby}
  \item{IntLeadership}{interest in leadership}
  \item{IntMgmt}{interest in management}
  \item{HSAnsbach1}{Ansbach first choice}
  \item{AlternativeHS}{name of alternative university}
  \item{AlternativerStg}{name of alternative program}
  \item{DistanzElternhaus}{distance of residence to residence of parents}
  \item{Note}{expected grade}
}
}
\usage{
studierende
}
\description{
The dataset was created from student data.
}
\keyword{datasets}
