% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{theme.own}
\alias{theme.own}
\title{Visualization theme}
\format{
List with 93 elements
\describe{
  \item{line}{line}
  \item{rect}{rect}
  \item{...}{...}
}
}
\usage{
theme.own
}
\description{
The theme defines the standards for your own organization.
}
\keyword{datasets}
