% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{theme.standard}
\alias{theme.standard}
\title{Visualization theme}
\format{
List with 93 elements
\describe{
  \item{line}{line}
  \item{rect}{rect}
  \item{...}{...}
}
}
\usage{
theme.standard
}
\description{
The theme defines the standards when r is started.
}
\keyword{datasets}
