% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariablesSelector.R
\name{VariableSelector}
\alias{VariableSelector}
\alias{VariableRanker}
\title{Ranks of importance variables}
\usage{
VariableRanker(formula, data, based = "gainratio", ...)
}
\arguments{
\item{formula}{A character object that contain the name of response variable about which a researcher is asking a question. Default value: "Y"}

\item{data}{the data frame that contains the variables specified in \code{formula}.}

\item{based}{Methodology used to rank variables. The options available are informationgain, gainratio and symmetrical.uncertainty.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A MLA object of subclass Var-Rank
}
\description{
A Ranker of variables
}
\examples{
## Load a Dataset
data(EGATUR)
VariableRanker(formula=GastoTotalD~pais+aloja+motivo,EGATUR)

}
