% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.eval.r
\name{ml.eval}
\alias{ml.eval}
\title{Function to evaluate a machine learning model the correlation for numeric variables}
\usage{
ml.eval(modell, daten)
}
\arguments{
\item{modell}{is the name of a machine learning model}

\item{daten}{is a dataframe with data}
}
\value{
model name, dataframe name, rmse & rsq
}
\description{
Function to evaluate a machine learning model the correlation for numeric variables
}
\keyword{evaluation}
\keyword{learning,}
\keyword{machine}
\keyword{modeling,}
